+++ src/components/LabResults.js
@@ -1,0 +1,11 @@
+
+
+import CardTitle from "@/components/CardTitle";
+
+const LabResults = () => (
+    <div class="flex flex-col bg-white rounded-xl ">
+        <CardTitle  title='Lab Results'></CardTitle>
+    </div>
+);
+
+export default LabResults
