+++ src/data/Patients.js
@@ -15,16 +15,16 @@ import avatar12 from "@public/avatar12.png";
 
 
 export const patients = [
-    { avatar: avatar1, name: 'Emily Williams', gender: 'Female', age: 18, active:false},
-    { avatar: avatar2, name: 'Ryan Johnson', gender: 'Male', age: 45, active:false},
-    { avatar: avatar3, name: 'Brandon Mitchell', gender: 'Male', age: 36, active:false},
-    { avatar: avatar4, name: 'Jessica Taylor', gender: 'Female', age: 28, active:true},
-    { avatar: avatar5, name: 'Samantha Johnson', gender: 'Female', age: 56, active:false},
-    { avatar: avatar6, name: 'Ashley Martinez', gender: 'Female', age: 54, active:false},
-    { avatar: avatar7, name: 'Olivia Brown', gender: 'Female', age: 32, active:false},
-    { avatar: avatar8, name: 'Tyler Davis', gender: 'Male', age: 19, active:false},
-    { avatar: avatar9, name: 'Kevin Anderson', gender: 'Male', age: 30, active:false},
-    { avatar: avatar10, name: 'Dylan Thompson', gender: 'Male', age: 36, active:false},
-    { avatar: avatar11, name: 'Nathan Evans', gender: 'Male', age: 58, active:false},
-    { avatar: avatar12, name: 'Mike Nolan', gender: 'Male', age: 31, active:false},
+    { id: 1, avatar: avatar1, name: 'Emily Williams', gender: 'Female', age: 18, active:false},
+    { id: 2, avatar: avatar2, name: 'Ryan Johnson', gender: 'Male', age: 45, active:false},
+    { id: 3, avatar: avatar3, name: 'Brandon Mitchell', gender: 'Male', age: 36, active:false},
+    { id: 4, avatar: avatar4, name: 'Jessica Taylor', gender: 'Female', age: 28, active:true},
+    { id: 5, avatar: avatar5, name: 'Samantha Johnson', gender: 'Female', age: 56, active:false},
+    { id: 6, avatar: avatar6, name: 'Ashley Martinez', gender: 'Female', age: 54, active:false},
+    { id: 7, avatar: avatar7, name: 'Olivia Brown', gender: 'Female', age: 32, active:false},
+    { id: 8, avatar: avatar8, name: 'Tyler Davis', gender: 'Male', age: 19, active:false},
+    { id: 9, avatar: avatar9, name: 'Kevin Anderson', gender: 'Male', age: 30, active:false},
+    { id: 10, avatar: avatar10, name: 'Dylan Thompson', gender: 'Male', age: 36, active:false},
+    { id: 11, avatar: avatar11, name: 'Nathan Evans', gender: 'Male', age: 58, active:false},
+    { id: 12, avatar: avatar12, name: 'Mike Nolan', gender: 'Male', age: 31, active:false},
 ];
